+++ sbroad/sbroad-core/src/frontend/sql/type_system.rs
@@ -447,6 +447,11 @@ fn default_type_system() -> TypeSystem {
         Function::new_operator("||", [Text, Text], Text),
         // Functions.
         Function::new_scalar(
+            get_real_function_name("version").expect("shouldn't fail"),
+            [],
+            Text,
+        ),
+        Function::new_scalar(
             // TODO:
             // Deprecated, remove in the future version.
             // Consider using `pico_instance_uuid` instead.